Refactor gradle build files.
This applies the methodology of blessed dependencies. Blessed
dependencies are dependencies which are untouchable, even when there is
a dep. conflict from transitive libraries. This is vital when dealing
with release builds, and automated testing since we can garuntee that
any vital dependency will be as-is, and that our testing environment
will be consistent.

epoxy/issues/49