Refactor and cleanup of gradle files.
This moves the java version into blessedDeps.gradle since we do not need
the flexibility with the java versioning. This also places our blessed
deps import into our base buildscript where it applys it to all
subprojects. We do not need to use `apply from:....` in every build file
now.