Gradle refactor issue#49 (#51)
* Refactor gradle build files.

This applies the methodology of blessed dependencies. Blessed
dependencies are dependencies which are untouchable, even when there is
a dep. conflict from transitive libraries. This is vital when dealing
with release builds, and automated testing since we can garuntee that
any vital dependency will be as-is, and that our testing environment
will be consistent.

epoxy/issues/49

* Rebuild gradle wrapper

I decided to recreate the gradle wrapper since there was an odd bug in
the build script where we are running into an issue where we aren't
finding the support annotations library on our path (?) Seems to
conflict with IDEAs annotation processor.

* Refactor and cleanup of gradle files.

This moves the java version into blessedDeps.gradle since we do not need
the flexibility with the java versioning. This also places our blessed
deps import into our base buildscript where it applys it to all
subprojects. We do not need to use `apply from:....` in every build file
now.