Block item click callback during item layout.
	Adding a RecyclerView.NO_POSITION != viewHolder.getAdapterPosition()
check inside of the click listener wrapper ensures we do not call the
click listener when the item is going through a layout.
	There was a need to make a change to the
ControllerLifecycleHelper to ensure that click still passed as the
adapter was always supplying `NO_POSITION` for unit tests. The simplest
way to get the basic functionality was to spy the viewHolder, and stub
this method.

ISSUE #293