proposal - instead of linking the view pool to activities, could we link to a lifecycle owner?
The problem is with Hilt, there is a memory leak. The memory leak is because Epoxy link the view pool to the activity however, when a view get inflated in a Hilt fragment, the context is a wrapper and hold a reference on the fragment. The fragment leak.

So why Epoxy link the view pool arbitrary to the `Activity`. Instead, could we use the view's context and find the closest `LifecycleOwner`. Use the `LifecycleOwner`'s lifecycle ti decide when to clear the pool.

However this change won't be enough to fix the memory leak we have with Hilt. We will also need this change in Hilt :