Add unbind extension to ItemViewBindingEpoxyHolder
In epoxy-sample, we show an example of using `unbind` to avoid memory leaks. Add an extension function for unbind to ItemViewBindingEpoxyHolder so that view binding can be used to clear listeners too.
The `unbind` extension function can be an optional override since not all ViewHolders will have listeners/resource allocation.

Also add example usage to kotlin-sample.