Dispose composition of `ComposeView` when unbinding `ComposeEpoxyModel`
Right now the composition of `ComposeView` in `ComposeEpoxyModel` isn't properly disposed. This leads to issues where e.g. when navigating to away from the screen where the `ComposeEpoxyModel` is rendered, and then coming back, the model isn't rendered anymore.

This PR fixes this by calling `ComposeView.disposeComposition()` when unbinding the `ComposeView` from the model.